first naive implementation of UI connections

+++ src/godot-src/GameWorld/Components/TorpedoLauncherComponent/torpedo_launcher_component.gd
@@ -1,10 +1,17 @@
 extends Node2D
 
+class_name PlayerTorpedoLauncher
+
 var torpedo_ref: WeakRef
 
 ###############################################################################
 # Builtin functions                                                           #
 ###############################################################################
+
+
+func _ready():
+	InputMediator.interface_state_changed.connect(unload_torpedo)
+
 
 func _process(delta):
 	if !torpedo_ref:
@@ -26,12 +33,16 @@ func _draw():
 		draw_line(Vector2.ZERO, mouse_position, Color.WHITE, 4.0, true)
 
 
-func _unhandled_input(event: InputEvent):
+func _input(event):
 	if event is InputEventMouseButton and event.button_mask == MOUSE_BUTTON_LEFT:
+		if !torpedo_ref:
+			return
+		# launch missile
 		if torpedo_ref.get_ref():
 			torpedo_ref.get_ref().delete()
 		torpedo_ref = null
 		queue_redraw()
+
 
 ###############################################################################
 # Public functions                                                            #
@@ -41,8 +52,11 @@ func load_torpedo(torpedo: InventoryItem):
 	torpedo_ref = weakref(torpedo)
 
 
-func unload_torpedo():
+func unload_torpedo(op_args: Variant = null):
+	if op_args == GameInputMediator.InterfaceState.MissileLaunch:
+		return
 	torpedo_ref = null
+	queue_redraw()
 
 
 ###############################################################################
